+++ src/components/Command.vue
@@ -141,6 +141,7 @@ const handleEnter = async () => {
       break;
     case "copy-path":
       await commands.page.copyPath();
+      hideMainUI();
       break;
     case "open-in-vscode":
       await commands.page.openInVSCode();